requirements.txt
azure-functions
azure-storage-blob
azure-cognitiveservices-vision-computervision
msrest
